Add clang-tidy rules and fix all the things
Note: something in here appears to break linking in release build

+++ icespider/core/exceptions.h
@@ -8,12 +8,12 @@
 	class DLL_PUBLIC Name : public ::IceSpider::HttpException { \
 		public: \
 			Name(); \
-			static const int code; \
+			static const short code; \
 			static const std::string message; \
 	}
 #define DefineHttpEx(Name, Code, Message) \
 	Name::Name() : ::IceSpider::HttpException(__FILE__, __LINE__, code, message) { } \
-	const int Name::code(Code); \
+	const short Name::code(Code); \
 	const std::string Name::message(Message);
 
 namespace IceSpider {
